Conflicts:
	frameworks/js-bindings/bindings/manual/ScriptingCore.cpp
	frameworks/js-bindings/bindings/script/jsb.js
	frameworks/js-bindings/cocos2d-x

+++ frameworks/js-bindings/bindings/manual/cocos2d_specifics.cpp
@@ -1374,7 +1374,7 @@ bool js_cocos2dx_CCNode_unscheduleAllSelectors(JSContext *cx, uint32_t argc, jsv
 		// For details to reproduce it, please refer to SchedulerTest/SchedulerUpdate.
         if(! arr) return true;
         JSScheduleWrapper* wrapper = NULL;
-        for(unsigned int i = 0; i < arr->count(); ++i) {
+        for(ssize_t i = 0; i < arr->count(); ++i) {
             wrapper = (JSScheduleWrapper*)arr->getObjectAtIndex(i);
             if(wrapper) {
                 cobj->getScheduler()->unscheduleAllForTarget(wrapper);
@@ -1636,7 +1636,7 @@ bool js_cocos2dx_CCNode_unscheduleUpdate(JSContext *cx, uint32_t argc, jsval *vp
             if(! arr) return true;
             
             JSScheduleWrapper* wrapper = NULL;
-            for(unsigned int i = 0; i < arr->count(); ++i) {
+            for(ssize_t i = 0; i < arr->count(); ++i) {
                 wrapper = (JSScheduleWrapper*)arr->getObjectAtIndex(i);
                 if(wrapper && wrapper->isUpdateSchedule()) {
                     cobj->getScheduler()->unscheduleUpdate(wrapper);
@@ -1734,7 +1734,7 @@ bool js_cocos2dx_CCScheduler_unscheduleAllSelectorsForTarget(JSContext *cx, uint
             if(! arr) return true;
             
             JSScheduleWrapper* wrapper = NULL;
-            for(unsigned int i = 0; i < arr->count(); ++i) {
+            for(ssize_t i = 0; i < arr->count(); ++i) {
                 wrapper = (JSScheduleWrapper*)arr->getObjectAtIndex(i);
                 if(wrapper) {
                     cobj->unscheduleAllForTarget(wrapper);
@@ -1851,7 +1851,7 @@ bool js_CCScheduler_unscheduleUpdateForTarget(JSContext *cx, uint32_t argc, jsva
             if(! arr) return true;
             
             JSScheduleWrapper* wrapper = NULL;
-            for(unsigned int i = 0; i < arr->count(); ++i) {
+            for(ssize_t i = 0; i < arr->count(); ++i) {
                 wrapper = (JSScheduleWrapper*)arr->getObjectAtIndex(i);
                 if(wrapper && wrapper->isUpdateSchedule()) {
                     cobj->unscheduleUpdate(wrapper);
@@ -1971,7 +1971,7 @@ bool js_CCScheduler_unscheduleCallbackForTarget(JSContext *cx, uint32_t argc, js
             if(! arr) return true;
             
             JSScheduleWrapper* wrapper = NULL;
-            for(unsigned int i = 0; i < arr->count(); ++i) {
+            for(ssize_t i = 0; i < arr->count(); ++i) {
                 wrapper = (JSScheduleWrapper*)arr->getObjectAtIndex(i);
                 if(wrapper && wrapper->getJSCallbackFunc() == argv[1]) {
                     cobj->unschedule(schedule_selector(JSScheduleWrapper::scheduleFunc), wrapper);
@@ -2040,7 +2040,7 @@ bool js_cocos2dx_CCScheduler_pauseTarget(JSContext *cx, uint32_t argc, jsval *vp
 			JSObject *tmpObj = JSVAL_TO_OBJECT(argv[0]);
             __Array *arr = JSScheduleWrapper::getTargetForJSObject(tmpObj);
             if(! arr) return true;
-            for(unsigned int i = 0; i < arr->count(); ++i) {
+            for(ssize_t i = 0; i < arr->count(); ++i) {
                 if(arr->getObjectAtIndex(i)) {
                     sched->pauseTarget(arr->getObjectAtIndex(i));
                 }
@@ -2066,7 +2066,7 @@ bool js_cocos2dx_CCScheduler_resumeTarget(JSContext *cx, uint32_t argc, jsval *v
 			JSObject *tmpObj = JSVAL_TO_OBJECT(argv[0]);
             auto arr = JSScheduleWrapper::getTargetForJSObject(tmpObj);
             if(! arr) return true;
-            for(unsigned int i = 0; i < arr->count(); ++i) {
+            for(ssize_t i = 0; i < arr->count(); ++i) {
                 if(arr->getObjectAtIndex(i)) {
                     sched->resumeTarget(arr->getObjectAtIndex(i));
                 }
@@ -2093,7 +2093,7 @@ bool js_cocos2dx_CCScheduler_isTargetPaused(JSContext *cx, uint32_t argc, jsval 
 			JSObject *tmpObj = JSVAL_TO_OBJECT(argv[0]);
             __Array *arr = JSScheduleWrapper::getTargetForJSObject(tmpObj);
             if(! arr) return true;
-            for(unsigned int i = 0; i < arr->count(); ++i) {
+            for(ssize_t i = 0; i < arr->count(); ++i) {
                 if(arr->getObjectAtIndex(i)) {
                     ret = cobj->isTargetPaused(arr->getObjectAtIndex(i)) ? true : false;
                     // break directly since all targets have the same `pause` status.
